refactor dcp manager
