+++ vllm/v1/attention/backends/flashinfer.py
@@ -74,6 +74,7 @@ from vllm.v1.attention.backends.utils import (
 )
 from vllm.v1.attention.ops.common import cp_lse_ag_out_rs
 from vllm.v1.attention.ops.dcp_alltoall import dcp_a2a_lse_reduce
+from vllm.v1.attention.ops.dcp_utils import FlashInferDCPManager
 from vllm.v1.attention.ops.merge_attn_states import merge_attn_states
 from vllm.v1.kv_cache_interface import (
     AttentionSpec,
@@ -82,7 +83,6 @@ from vllm.v1.kv_cache_interface import (
     UniformTypeKVCacheSpecs,
 )
 from vllm.v1.utils import CpuGpuBuffer
-from vllm.v1.worker.ubatching import dbo_current_ubatch_id
 
 FLASHINFER_WORKSPACE_BUFFER_SIZE_BATCH_INVARIANT = 2048 * 1024 * 1024
 FLASHINFER_PREFILL_WORKSPACE_BYTES_PER_ELEM = 16
@@ -93,72 +93,6 @@ FP4_DTYPE = torch.uint8
 logger = init_logger(__name__)
 
 trtllm_workspace_buffer = None
-
-
-class _FlashInferDCPManager:
-    """Own persistent buffers and collectives for FlashInfer DCP decode."""
-
-    def __init__(
-        self,
-        vllm_config: VllmConfig,
-        num_heads: int,
-        head_size: int,
-        dtype: torch.dtype,
-    ) -> None:
-        parallel_config = vllm_config.parallel_config
-        scheduler_config = vllm_config.scheduler_config
-        speculative_config = vllm_config.speculative_config
-
-        self.world_size = parallel_config.decode_context_parallel_size
-        self.max_seq_len = scheduler_config.max_num_seqs
-        num_spec_tokens = (
-            speculative_config.num_speculative_tokens
-            if speculative_config is not None
-            else 0
-        )
-        self.max_num_tokens = min(
-            scheduler_config.max_num_batched_tokens,
-            self.max_seq_len * (1 + num_spec_tokens),
-        )
-
-        num_heads_in_dcp = num_heads * self.world_size
-        num_ubatches = max(parallel_config.num_ubatches, 1)
-        self.output = torch.empty(
-            (
-                num_ubatches,
-                self.max_num_tokens,
-                num_heads_in_dcp,
-                head_size,
-            ),
-            dtype=dtype,
-            device="cuda",
-        )
-        self.lse = torch.empty(
-            (num_ubatches, self.max_num_tokens, num_heads_in_dcp),
-            dtype=torch.float32,
-            device="cuda",
-        )
-
-        combine_fn = (
-            dcp_a2a_lse_reduce
-            if parallel_config.dcp_comm_backend == "a2a"
-            else cp_lse_ag_out_rs
-        )
-        self.combine = partial(combine_fn, is_lse_base_on_e=False)
-
-    def get_decode_buffers(
-        self, num_decode_tokens: int
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        if num_decode_tokens > self.max_num_tokens:
-            raise RuntimeError(
-                "FlashInfer DCP decode batch exceeds the persistent buffer: "
-                f"{num_decode_tokens} > {self.max_num_tokens}."
-            )
-        ubatch_id = dbo_current_ubatch_id()
-        return (
-            self.output[ubatch_id, :num_decode_tokens],
-            self.lse[ubatch_id, :num_decode_tokens],
-        )
 
 
 def _get_trtllm_workspace_buffer():
@@ -1830,7 +1764,7 @@ class FlashInferImpl(AttentionImpl):
             else 1
         )
         self.dcp_manager = (
-            _FlashInferDCPManager(
+            FlashInferDCPManager(
                 vllm_config,
                 num_heads,
                 head_size,
@@ -2320,9 +2254,7 @@ class FlashInferImpl(AttentionImpl):
                     output_tmp, lse = self.dcp_manager.get_decode_buffers(
                         num_decode_tokens
                     )
-                    decode_query = get_dcp_group().all_gather(
-                        decode_query.contiguous(), dim=-2
-                    )
+                    decode_query = self.dcp_manager.gather_query(decode_query)
                     decode_wrapper.run(
                         decode_query,
                         kv_cache_for_fi,
@@ -2338,7 +2270,6 @@ class FlashInferImpl(AttentionImpl):
                     output[:num_decode_tokens] = self.dcp_manager.combine(
                         output_tmp,
                         lse,
-                        get_dcp_group(),
                     )
                 else:
                     decode_wrapper.run(
